Add files via upload
This Query combines multiple SELECT statements using UNION ALL to classify employees based on salary and age. It labels employees with salaries over $65,000 as "expensive", males over 40 as "OLD MAN", and females over 40 as "OLD lady". The result is a unified list of employees with these labels. A temporary table OLD_and_expensive is created to store this data, and the final result is ordered by first name for easy viewing of the categorized employees.